+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/Common/FloatingActions/FloatingActionPrivacyMutePresenter.cs
@@ -8,7 +8,6 @@ using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.Common.FloatingActions;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.Common.FloatingActions;
-using ICD.Profound.ConnectPRO.Utils;
 
 namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common.FloatingActions
 {
@@ -79,7 +78,7 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common.Floatin
 		/// </summary>
 		private void UpdateVisibility()
 		{
-			bool show = Room != null && ConferenceOverrideUtils.ConferenceActionsAvailable(Room, eInCall.Audio);
+			bool show = Room != null && Room.ConferenceActionsAvailable(eInCall.Audio);
 			ShowView(show);
 		}